Added new base factory module to clean up and separate components Refactored factories and general tidy up Fixing tests
