+++ spec/services/cms/dynamic_components/card_wrapper_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 RSpec.describe Cms::DynamicComponents::CardWrapper do
   before do
     card_section_data = Cms::Mocks::NumericCardSection.generate_data
-    @card_wrapper = Cms::Providers::Strapi::Factories::ComponentFactory.to_card_wrapper(card_section_data, card_section_data[:numericCards])
+    @card_wrapper = Cms::Providers::Strapi::Factories::BlocksFactory.to_card_wrapper(card_section_data, card_section_data[:numericCards])
   end
 
   it "should render as CmsCardWrapperComponent" do
