+++ spec/services/cms/dynamic_components/full_width_banner_spec.rb
@@ -2,7 +2,7 @@ require "rails_helper"
 
 RSpec.describe Cms::DynamicComponents::FullWidthBanner do
   before do
-    @banner = Cms::Providers::Strapi::Factories::ComponentFactory.to_full_width_banner(Cms::Mocks::FullWidthBanner.generate_data)
+    @banner = Cms::Providers::Strapi::Factories::ComponentFactory.process_component(Cms::Mocks::FullWidthBanner.generate_raw_data)
   end
 
   it "should render as FullWidthBannerComponent" do
