+++ spec/services/cms/dynamic_components/horizontal_card_spec.rb
@@ -2,7 +2,7 @@ require "rails_helper"
 
 RSpec.describe Cms::DynamicComponents::HorizontalCard do
   before do
-    @card = Cms::Providers::Strapi::Factories::ComponentFactory.to_horizontal_card(Cms::Mocks::HorizontalCard.generate_data)
+    @card = Cms::Providers::Strapi::Factories::ComponentFactory.process_component(Cms::Mocks::HorizontalCard.generate_raw_data)
   end
 
   it "should render as HorizontalCardComponent" do
