+++ spec/services/cms/dynamic_components/question_and_answer_spec.rb
@@ -3,8 +3,8 @@ require "rails_helper"
 RSpec.describe Cms::DynamicComponents::QuestionAndAnswer do
   before do
     stub_strapi_aside_section("test-aside")
-    data = Cms::Mocks::QuestionAndAnswer.generate_data(asideSlugs: Cms::Mocks::AsideSection.generate_aside_list(aside_slugs: ["test-aside"]))
-    @qa_component = Cms::Providers::Strapi::Factories::ComponentFactory.to_question_and_answer(data)
+    data = Cms::Mocks::QuestionAndAnswer.generate_raw_data(asideSlugs: Cms::Mocks::AsideSection.generate_aside_list(aside_slugs: ["test-aside"]))
+    @qa_component = Cms::Providers::Strapi::Factories::ComponentFactory.process_component(data)
   end
 
   it "should render as CmsQuestionAndAnswerComponent" do
