+++ spec/services/cms/dynamic_components/testimonial_row_spec.rb
@@ -2,7 +2,7 @@ require "rails_helper"
 
 RSpec.describe Cms::DynamicComponents::TestimonialRow do
   before do
-    @testimonial_row = Cms::Providers::Strapi::Factories::ComponentFactory.to_testimonial_row(Cms::Mocks::TestimonialRow.generate_data)
+    @testimonial_row = Cms::Providers::Strapi::Factories::ComponentFactory.process_component(Cms::Mocks::TestimonialRow.generate_raw_data)
   end
 
   it "should render as CmsTestimonialRowComponent" do
